Split tests into specs
