+++ swiftzTests/ShapeExample.swift
@@ -6,6 +6,7 @@
 //  Copyright (c) 2014 Maxwell Swadling. All rights reserved.
 //
 
+import XCTest
 import swiftz
 
 // shape example for SYB
@@ -48,5 +49,36 @@ func ==(lhs: Shape, rhs: Shape) -> Bool {
 		return true
 	default: 
 		return false
+	}
+}
+
+class GenericsSpec : XCTestCase {
+	func testGenericsSYB() {
+		let b = Shape.Plane(2)
+		let b2 = Shape.fromRep(b.toRep())
+		XCTAssert(b == b2!)
+
+		// not sure why you would use SYB at the moment...
+		// without some kind of extendable generic dispatch, it isn't very useful.
+		let gJSON : Data -> JSONValue = { d in
+			var r = Dictionary<String, JSONValue>()
+			for (n, vs) in d.vals {
+				switch vs {
+				case let x as Int:
+					r[n] = JSONValue.JSONNumber(Double(x))
+				case let x as String:
+					r[n] = JSONValue.JSONString(x)
+				case let x as Bool:
+					r[n] = JSONValue.JSONBool(x)
+				case let x as Double:
+					r[n] = JSONValue.JSONNumber(x)
+				default:
+					r[n] = JSONValue.JSONNull()
+				}
+			}
+			return .JSONObject(r)
+		}
+
+		XCTAssert(gJSON(b.toRep()) == .JSONObject(["wingspan" : .JSONNumber(2)]))
 	}
 }
